Initial setup for move to canvasapi open source package
