+++ rudaux/tasks.py
@@ -10,7 +10,7 @@ def get_class_from_string(s):
 
 # @task
 def get_learning_management_system(settings, group_name):
-    LMS = get_class_from_string(settings.lms_classes[group_name])
+    LMS = get_class_from_string(settings['lms_classes'][group_name])
     if not issubclass(LMS, LearningManagementSystem):
         raise ValueError
     lms = LMS.parse_obj(settings)
@@ -30,7 +30,7 @@ def get_grading_system(settings, group_name):
 
 # @task
 def get_submission_system(settings, group_name):
-    SubS = get_class_from_string(settings.ss_classes[group_name])
+    SubS = get_class_from_string(settings['ss_classes'][group_name])
     if not issubclass(SubS, SubmissionSystem):
         raise ValueError
     subs = SubS.parse_obj(settings)
